MacGui: Never disable the chapter markers checkbox. Patch by Rodeo ... Thanks!
- Discussed here https://reviews.handbrake.fr/r/58/
This allows users to enable/disable chapter markers regardless of whether the source has chapters, which is very useful when saving a preset.

This will not cause chapter markers to be inserted if we only have one chapter, as this is sanitized twice:
 - once in macosx/Controller.m, in createQueueFileItem
 - once in libhb/work.c, in do_job()

In autoSetM4vExtension, update the condition for chapter markers being on or off, since we can no longer rely on the checkbox being enabled/disabled.

git-svn-id: svn://localhost/HandBrake/trunk@3888 b64f7644-9d1e-0410-96f1-a4d463321fa5