Fix a problem with BD audio detection.
Each clip of a BD are allowed to have different audios if the clip
does not have a seamless connection to the previous clip.  Most titles
are a series of seamless clips that all have the exact same audio.  But
I found some that have a final non-seamless clip that has completely 
different audios and broke the old algorithm.  

New algorithm, look at each clip and count the number of other clips have
the same audio.  Use the clip that has the most matches.


git-svn-id: svn://localhost/HandBrake/trunk@3918 b64f7644-9d1e-0410-96f1-a4d463321fa5