Add support for TrueHD and DTS-HD from BD sources TrueHD and DTS-HD now show up in the audio list along side their AC-3 and DTS counterparts.
Note that currently the DTS-HD decoder we are using (ffmpeg) discards
the HD portion of the stream and onle decodes the DTS core portion.  So
there is no advantage yet to using the DTS-HD stream.  In the future
I would like to add DTS-HD passthru support and hopefully ffmpeg will
improve their DTS-HD decoder.


git-svn-id: svn://localhost/HandBrake/trunk@3950 b64f7644-9d1e-0410-96f1-a4d463321fa5