Fix parsing of SSA layer field
SSA subtitles have an empty layer field (bare ',').  The
scanf format specifier "%*128[^,]" will not match on a
bare ','.  There must be at least one non ',' character in
the match.  So the format specifier is placed directly next
to the 'Dialogue:' string so that the next expected ' '
after the ':' will be the character it matches on when
there is no layer field.

git-svn-id: svn://localhost/HandBrake/trunk@4244 b64f7644-9d1e-0410-96f1-a4d463321fa5