handle PCR discontinuities better
The PCR comparison we were doing could cause us to drop all the
timestamps for an audio stream when the audio packets are very widely
spaced.  Now, it detects PCR changes explicitely and only drops the
timestamps if a discontinuity PCR prior to the packets current PCR has
already been sent to the reader.

git-svn-id: svn://localhost/HandBrake/trunk@4263 b64f7644-9d1e-0410-96f1-a4d463321fa5