fix crash during 2 pass encoding TS or PS source
stream.c cached information that it probed during the first time it
opened any source.  Then later it would re-use that cached data.  I
was prematurely deleting the cached data during 2 pass encodes.

The problem is that there is no way to know when the cached data is no
longer needed.  You could have a thousand items in the queue all using
the same source, or you could have only 1.  So you either have to (a)
keep the cached data indefinitely, or (b) you have to be able to handle
the case where scanned cached data is flushed before you start an
encode. (a) is poor design.  And if you choose (b) you might as well
eliminate the cache all together.  It doesn't really save any time and
only complicates the code.

In summary, the cache is gone.

git-svn-id: svn://localhost/HandBrake/trunk@4286 b64f7644-9d1e-0410-96f1-a4d463321fa5