detect progressive upsampling in vfr
progressive 23.976 fps content that has upsampled to progressive 59.94
fps can result in juddery output if you don't drop the "right" frames
while downsampling back to 23.976.  This adds a test that computes the
sum of squared errors between 2 candidate frames to drop and makes a
more intelligent decision based on this metric.

git-svn-id: svn://localhost/HandBrake/trunk@4297 b64f7644-9d1e-0410-96f1-a4d463321fa5