Fix interleaving of subtitles in muxer
Specifically, this affects closed captions which are pushed into the
pipeline very late.  CC's are dribbled in 2 bytes per video frame.
The entire CC that we put into a subtitle frame isn't available till
we see the signal for the end of the CC.  This can be several seconds
after it began.  So I have established a minimum buffer size that the
muxer accumulates before it outputs any data.  This allows CCs to get
into the pipeline before the muxer outputs the interleave interval that
they belong in.

git-svn-id: svn://localhost/HandBrake/trunk@4351 b64f7644-9d1e-0410-96f1-a4d463321fa5