libhb: Fix problem with positions of burned-in subs When cropping and/or scaling, the position of burned-in SSA subs was broken. Also, when scaling, the position of all other burned-in subs was broken.
Part of this fix is to revert a bit of filter initialization changes that
were made in this commit https://trac.handbrake.fr/changeset/4605

The other part of the fix is to add cropping parameters to the initialization
of the subtitle rendering filter.  This filter needs the *original* title
dimensions and the crop values in order to compute the positions properly.
The changes that I am reverting gave it the scaled job dimensions and the
crop values.  This was wrong in so many ways it's embarassing :p


git-svn-id: svn://localhost/HandBrake/trunk@4642 b64f7644-9d1e-0410-96f1-a4d463321fa5