libhb: Fix include with the system's libav.
This allows one to avoid including a header that is declared by upstream to
be an internal header, that is, one that should not be `#include`'d. For
further information, see [the discussion in Debian's BTS][0].

[0]: bugs.debian.org/672561#25

In particular, this is needed when one wants to use a system's ffmpeg/libav,
instead of downloading it from handbrake's site.

Thanks to Fabian Greffrath for the more elegant solution than the "brute
force" that I employed earlier.

Reviewed-by: Rogério Theodoro de Brito <rbrito@ime.usp.br>
Signed-off-by: Rogério Theodoro de Brito <rbrito@ime.usp.br>